refactor: (#5) pages FSD 구조 정리
