+++ apps/web/app/loading.tsx
@@ -1,4 +1,4 @@
-import { LoadingPage } from "@/pages";
+import { LoadingPage } from "@/pages/loading";
 
 export default function Loading() {
   return <LoadingPage />;
