+++ apps/web/app/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginPage } from "@/pages/LoginPage";
+import { LoginPage } from "@/pages/login";
 
 export default function Page() {
   return <LoginPage />;
