+++ apps/web/app/not-found.tsx
@@ -1,4 +1,4 @@
-import { NotFoundPage } from "@/pages";
+import { NotFoundPage } from "@/pages/not-found";
 
 export default function NotFound() {
   return <NotFoundPage />;
