+++ apps/web/app/page.tsx
@@ -1,4 +1,4 @@
-import { RouteIndexPage } from "@/pages";
+import { RouteIndexPage } from "@/pages/route-index";
 
 export default function Page() {
   return <RouteIndexPage />;
