+++ apps/web/app/preview/error/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ErrorPage } from "@/pages";
+import { ErrorPage } from "@/pages/error";
 
 export default function PreviewErrorPage() {
   return (
