+++ apps/web/app/preview/loading/page.tsx
@@ -1,4 +1,4 @@
-import { LoadingPage } from "@/pages";
+import { LoadingPage } from "@/pages/loading";
 
 export default function PreviewLoadingPage() {
   return <LoadingPage />;
