+++ apps/web/app/preview/not-found/page.tsx
@@ -1,4 +1,4 @@
-import { NotFoundPage } from "@/pages";
+import { NotFoundPage } from "@/pages/not-found";
 
 export default function PreviewNotFoundPage() {
   return <NotFoundPage />;
