+++ apps/web/src/pages/error/index.ts
@@ -1,0 +1,2 @@
+export { ErrorPage } from "./ui/ErrorPage";
+export type { ErrorPageProps } from "./ui/ErrorPage";
