--- apps/web/src/pages/index.ts
@@ -1,5 +1,0 @@
-export { LoginPage } from "./LoginPage";
-export { LoadingPage } from "./LoadingPage";
-export { ErrorPage } from "./ErrorPage";
-export { NotFoundPage } from "./NotFoundPage";
-export { RouteIndexPage } from "./RouteIndexPage";
