+++ apps/web/src/pages/loading/index.ts
@@ -1,0 +1,1 @@
+export { LoadingPage } from "./ui/LoadingPage";
