+++ apps/web/src/pages/route-index/index.ts
@@ -1,0 +1,1 @@
+export { RouteIndexPage } from "./ui/RouteIndexPage";
